Stem attribute in cc_{binary|library}
By default, we use the name of the target as the name of the output
.so (or binary). However, this can be overriden using the `stem`
property. This CL implements this in our Bazel rule.

With `stem`, we can run into collisions since two targets in the same
package can use the same `stem`. To avoid this, the top-level .so file
is generated in a subdir specific to a target

Update existing b unit tests to use the updated location

Bug: 240563612
Test: b unit tests
Test: built //bionic/libc:libc and //bioinc/libc:libc_hwasan and
verified that both generate a file named libc.so

Change-Id: I34762ef9ae1ec601c87144310f9132e6c8e81217