Fix build break with mixed_mode_toggle.sh.
The test was previously checking the exact path for an output in
bazel-out. r.android.com/2606989 changed the default output path to not embed
the product, which now breaks the test, which assumed aosp_arm64 was
part of it.

This change makes the test a little more resilient to upstream changes
like that.

Test: ran the test.
Change-Id: I200e9518479d00d9e55935ad19a826882c6de9ba