remove Java backend from aidl_interface
There is a transitive dependency chain from av-audio-types-aidl to
android.hardware.audio.effect-V1 through the Java aidl aspect. However
android.hardware.audio.effect-V1-java needs the min_sdk_version flag to
be set. Since the Java aidl rules are implemented via an aspect, this
flag has to be set on the aidl_library itself.

In lieu of implementing this right now at 5:45p on Friday, this CL just
disables the Java backend for aidl_interfaces and we can revisit the
design of the Java aidl rules to decide if it's worth maintaining the
aspect-based implementation alongside the cc aidl rules which emulate
the Soong implementation.

Bug: 285574832
Test: bp2build.sh
Change-Id: I759caa389e6075fb1da8ee92aa60e3afcf044352