Expand overriding support for android_app_certificate.
It's legal to create a brand new android_app_certificate and use
PRODUCT_CERTIFICATE_OVERRIDES to reference the new module. It's also
legal in Android.bp to name the module and the certificate
property (which is the extension-less basename of the certs to pick up)
completely differently.

This CL adds support in Bazel for the above. The main change is to
OverridingCertificateFiles to provide a list of Targets instead of
actual files, so that we can check the name of the target against the
PRODUCT_CERTIFICATE_OVERRIDES value. If it matches, then we read its
files.

The original android_certificate_directory approach still works.

Bug: 284074879
Change-Id: I496dacbcae67229d4e2cd346131c031dabd510d9