[cc_test] Overhaul cc_test.bzl and tradefed.bzl.
This is a fairly large cl, so please read this commit message before
proceeding.

There are three main goals I wanted to achieve with this overhaul:

1) Get 'b test //...' to work - that is, a flagless invocation,
regardless of whether a test is deviceless and/or needs a device (host
driven, or device driven). If a test is supported for both (e.g.
hello_world_test), then both tests will run in parallel. The unsupported
variant will be skipped automatically.

2) Get the --roboleaf-mode tradefed ruleset to be as similar as possible
with the ones --bazel-mode uses.

3) Stop exposing the tradefed-less tests as test executables. For the
purposes of migration, we will only scope all tests that are running
with the Tradefed harness.

To achieve these goals, I've tried to refactor the build/bazel tradefed
rules to be as similar as possible to the ones under tools/asuite, which
is used by --bazel-mode. I've also made use of transitions and
target_compatible_with to ensure that:

1) The tradefed test rule entry points ALWAYS set target_compatible_with
to os:linux. Depending on the test type, it transitions the test
dependencies to either host or device.
2) The test dependencies themselves respect host_supported in bp2build,
which translates to target_compatible_with. The combination of
target_compatible_with and the outgoing transition ensures that:

  - Tests that don't build for host will automatically skip the deviceless test invocation.
  - Tests that don't build for device will automatically skip the device test invocations

Test: atest-dev --roboleaf-mode=dev hello_world_test adbd_test adb_tls_connection_test adb_pairing_connection_test adb_pairing_auth_test adb_crypto_test --bazel-arg=--nocache_test_results
--bazel-arg=--config=bes_metrics
Test: Results invocation ID: 898b0016-d68e-4188-819b-50e2d41feeb9

Test: b test //system/... //packages/modules/... --build_tests_only --config=bes_metrics
Test: Results invocation ID: 4d815194-19d6-4b34-a5bf-728099eca2f6

Bug: 280726231
Bug: 288869073
Bug: 288069169
Change-Id: If7e17985ec77770d193186f5d7165e15c7504644