Default the main of underlying py_binary to the module name
The `main` attribute is used to specify the main entry point of the
application. If it's not specified, the target name will be used
instead. This file must also be listed in srcs.

But in the py_test, we add a suffix to the original module name
for the underlying py_binary target, which causes mismatch between
the main file name and the target name if the `main` is not
specified.

Test: b test //frameworks/base/api:merge_annotation_zips_test
Test: presubmit
Bug: 293225210
Change-Id: I2a9dcc71e18fa41d9b1a05a5388e1b07d02994d1