Update include paths for .protos generated in synthetic dir
Some genrules are used to modify the paths for includes in .cc files.
e.g. there might be a folder like
- foo.cc
- foo.proto
- Android.bp
And foo.cc might want to include the header as "bar/foo.pb.h".

To build such deps, this CL adds an additional include to the compile
action of foo.cc

(Another option was a source directory cleanup, but I did not pursue it
further because Bazel is more strict about .protos being in the same
pacakge. The implementation I had in mind for this option would put the
proto in a different dir/package)

Test: Added an integration test to mimic the genrules usages in internal
Bug: 293205700
Change-Id: I1eb315de9e14e1ca43532a81b09225d9f357e1b4