Handle .proto in srcs that imports another .proto in that src
For the following tree,
```
tmp
├── foo.proto
└── import_foo.proto # contains an `import "foo.proto"`
├── Android.bp # contains a cc_library with these protos
```
We currently pass both protos to aprotoc for compilation. We use
proto_source_root as -I (which starts with bazel-out/...), but pass
f.short_path of the proto files (which does *not* start with bazel-out).
This meant that during the protoc compilation, it ends up with 3 proto
files in its database (relative to the execution dir)
1. tmp/foo.proto
2. tmp/import_foo.proto
3. import_foo.proto # from the import statement

(2) and (3) are the same, so protoc raises an exception. This CL fixes
it by passing f.path to protoc instead. These paths are superstrings of
the -I paths, which allows protoc to correctly identify the proto files

This new format required me to make updates to two other places
- Remove the special handling of protos via genrules (aosp/2687994)
- An existing cc proto test. The expected paths change in the new
  format, and I would like to make them explicit in the assert,
  as per go/tott/338 ("dont put logic in tests")

Bug: 293514502
Bug: 293205700
Test: Added an integration test
Test: Built the internal module of b/293514502
Test: TH

Change-Id: Ic812136fe762b4c476487bb932996538b16b8244