Fix `b test //packages/modules/adb:adb_test` not picking up any test.
adb_test is tagged with 'no-remote'. The tag was propagated to the
test_suite, but not the actual tests themselves, so the test_suite
filtered the tests away, leaving with no tests to run.

To fix this, add the tags to the tests in the macro so the test_suite
and the tests have the same tags.

Fixes: 294514745
Fixes: 268186228
Test: presubmits
Test: b test //packages/modules/adb:adb_test
Change-Id: Iac0814f53ab027437626ad7caa96d1d921915a49