Fixes for python 3.11
Bazel previously had a bug where the source directory of the entrypoint
python script would be added to the path. They added PYTHONSAFEPATH
to the enviornment to fix it, which was introduced in python 3.11.

Now that python is being updated to 3.11, the fix is in, which breaks
some scripts that don't properly specify import folders.

Add the import folders where required.

Kindof unforunately, imports works transitively, so adding imports
was only necessary on the lowest level libraries, despite the behavior
being relied upon at a higher level. I didn't attempt to solve this
right now.

Bug: 278602456
Test: ./build/bazel/ci/bp2build.sh
Change-Id: I4e8e746270da27920bb25e78b7114b6c21a275f8