Don't make output directories when building apexes
Ninja gives this message: `ninja: outputs should be files, not
directories: ...`

This is because a directory's mtime will only change if files in it are
added / removed. But if the files change contents, or files in
subdirectories are added/removed, the directory won't be considered
"dirty" by ninja, and dependant actions won't be rerun.

Bug: 293609369
Test: m
Change-Id: I5bdb87310832841b08472e85ad14ba9475a5a182