Refactor api.bzl to depend explicitly on product variables.
After this change, all functions in api.bzl depend explicitly on the
set of 4 product variables that affect them, and the api_internal.bzl
file is deleted. This makes the API more uniform and the code more
readable (in my opinion). The `api` struct is still constructed from the
load time product variables, but `api_from_product` can be used as an
alternative that takes as input a struct of the 4 variables
constructed from, e.g., build settings from a rule.

Bug: 297356809
Test: Unit tests + CI
Change-Id: Ic22cb68af0f1f5e352dcc1fd7e2667e005ae6ada