Add rust toolchain for arm, x86, and x86_64
The base toolchains set `target_settings` to `base_toolchain_enabled` config_setting which is set in the base_transition attached to `with_base_transition` rule.

The order of the regsistered toolchains matter to ensure that when `base_toolchain_enabled` is true, the base toolchain is preferred over the top-level toolchain used to build end-user rust targets.

Test: b build //build/bazel/examples/rust:all --config=android
Bug: 295918553
Change-Id: I5b7925e0ede11a7d82881cd67eadac2ac484bb47